+++ sema4ai/tests/sema4ai_code_tests/test_vscode_integration.py
@@ -9,6 +9,17 @@ from typing import Dict, List, Optional
 from unittest import mock
 
 import pytest
+from sema4ai_code_tests.fixtures import RCC_TEMPLATE_NAMES, RccPatch
+from sema4ai_code_tests.protocols import IRobocorpLanguageServerClient
+from sema4ai_ls_core.basic import wait_for_condition
+from sema4ai_ls_core.callbacks import Callback
+from sema4ai_ls_core.ep_resolve_interpreter import (
+    DefaultInterpreterInfo,
+    IInterpreterInfo,
+)
+from sema4ai_ls_core.pluginmanager import PluginManager
+from sema4ai_ls_core.unittest_tools.cases_fixture import CasesFixture
+
 from sema4ai_code.inspector.common import (
     STATE_CLOSED,
     STATE_NOT_PICKING,
@@ -21,17 +32,6 @@ from sema4ai_code.protocols import (
     LocalPackageMetadataInfoDict,
     WorkspaceInfoDict,
 )
-from sema4ai_ls_core.basic import wait_for_condition
-from sema4ai_ls_core.callbacks import Callback
-from sema4ai_ls_core.ep_resolve_interpreter import (
-    DefaultInterpreterInfo,
-    IInterpreterInfo,
-)
-from sema4ai_ls_core.pluginmanager import PluginManager
-from sema4ai_ls_core.unittest_tools.cases_fixture import CasesFixture
-
-from sema4ai_code_tests.fixtures import RCC_TEMPLATE_NAMES, RccPatch
-from sema4ai_code_tests.protocols import IRobocorpLanguageServerClient
 
 log = logging.getLogger(__name__)
 
@@ -1072,10 +1072,9 @@ def test_hover_image_integration(
 ):
     import base64
 
+    from sema4ai_code_tests.fixtures import IMAGE_IN_BASE64
     from sema4ai_ls_core import uris
     from sema4ai_ls_core.workspace import Document
-
-    from sema4ai_code_tests.fixtures import IMAGE_IN_BASE64
 
     locators_json = tmpdir.join("locators.json")
     locators_json.write_text("", "utf-8")
@@ -1346,13 +1345,14 @@ def test_profile_import(
     datadir,
     disable_rcc_diagnostics,
 ):
+    from sema4ai_ls_core import uris
+
     from sema4ai_code.commands import (
         SEMA4AI_GET_PY_PI_BASE_URLS_INTERNAL,
         SEMA4AI_PROFILE_IMPORT_INTERNAL,
         SEMA4AI_PROFILE_LIST_INTERNAL,
         SEMA4AI_PROFILE_SWITCH_INTERNAL,
     )
-    from sema4ai_ls_core import uris
 
     result = language_server_initialized.execute_command(
         SEMA4AI_GET_PY_PI_BASE_URLS_INTERNAL,
@@ -1671,11 +1671,10 @@ def test_web_inspector_integrated(
     This test should be a reference spanning all the APIs that are available
     for the inspector webview to use.
     """
-    from sema4ai_ls_core import uris
-
     from sema4ai_code_tests.robocode_language_server_client import (
         RobocorpLanguageServerClient,
     )
+    from sema4ai_ls_core import uris
 
     cases.copy_to("robots", ws_root_path)
     ls_client: RobocorpLanguageServerClient = language_server_initialized
@@ -2301,7 +2300,7 @@ def _write_agent_spec_and_pack(
     )["result"]
 
 
-def test_create_and_pack_agent_package(
+def test_create_and_pack_and_import_agent_package(
     language_server_initialized,
     tmpdir,
     agent_cli_location: str,
@@ -2309,6 +2308,7 @@ def test_create_and_pack_agent_package(
     assert os.path.exists(agent_cli_location)
 
     import yaml
+
     from sema4ai_code import commands
 
     language_server = language_server_initialized
@@ -2358,7 +2358,8 @@ def test_create_and_pack_agent_package(
 
     assert result["success"]
     assert not result["message"]
-    assert os.path.exists(f"{target_directory}/agent-package.zip")
+    agent_package_zip = f"{target_directory}/agent-package.zip"
+    assert os.path.exists(agent_package_zip)
 
     # Incorrect agent spec configuration
     del agent_spec["agent-package"]["agents"][0]["description"]
@@ -2372,12 +2373,39 @@ def test_create_and_pack_agent_package(
         "Missing required entry: agent-package/agents/description." in result["message"]
     )
 
+    no_args_request = language_server.request(
+        {
+            "jsonrpc": "2.0",
+            "id": language_server.next_id(),
+            "method": "importAgentPackageFromZip",
+            "params": {"target_dir": "", "agent_package_zip": ""},
+        }
+    )["result"]
+    assert not no_args_request["success"]
+
+    target_directory = str(tmpdir.join("extract_to"))
+
+    ok_request = language_server.request(
+        {
+            "jsonrpc": "2.0",
+            "id": language_server.next_id(),
+            "method": "importAgentPackageFromZip",
+            "params": {
+                "target_dir": target_directory,
+                "agent_package_zip": agent_package_zip,
+            },
+        }
+    )["result"]
+    assert ok_request["success"]
+    assert (Path(target_directory) / "agent-spec.yaml").exists()
+
 
 def test_get_runbook_from_agent_spec(
     language_server_initialized,
     tmpdir,
 ) -> None:
     import yaml
+
     from sema4ai_code import commands
 
     language_server = language_server_initialized