+++ sema4ai/vscode-client/src/robocorpCommands.ts
@@ -136,3 +136,4 @@ export const SEMA4AI_PACK_AGENT_PACKAGE_INTERNAL = "sema4ai.packAgentPackage.int
 export const SEMA4AI_OPEN_RUNBOOK_TREE_SELECTION = "sema4ai.openRunbookTreeSelection";  // Edit Runbook
 export const SEMA4AI_GET_RUNBOOK_PATH_FROM_AGENT_SPEC_INTERNAL = "sema4ai.getRunbookPathFromAgentSpec.internal";  // Get Runbook Path from Agent Spec (internal)
 export const SEMA4AI_AGENT_PACKAGE_PUBLISH_TO_SEMA4_AI_STUDIO_APP = "sema4ai.agentPackagePublishToSema4AIStudioApp";  // Publish Agent Package to Sema4.ai Studio
+export const SEMA4AI_AGENT_PACKAGE_IMPORT = "sema4ai.agentPackageImport";  // Import Agent Package (Zip)